requirements.txt
streamlit==1.38.0
requests==2.32.3
python-dotenv==1.0.1
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4
pillow-heif==0.18.0

